pyproject.toml
[project]
name = "automa_ai"
version = "0.1.1"
description = "PNNL Autoflow-AI: Dynamic multi-agent system for building applications"
requires-python = ">=3.12"
authors = [{name = "Weili Xu", email = "weili.xu@pnnl.gov"}, {name = "Hanlong, Wan", email = "hanlong.wan@pnnl.gov"}]
license = "BSD-2-Clause"
readme = "README.md"
dependencies = [
    "a2a-sdk>=0.2.8",
    "asyncclick>=8.1.8",
    "black>=25.1.0",
    "bs4>=0.0.2",
    "chromadb>=1.0.13",
    "google-adk>=1.4.1",
    "google-genai>=1.21.1",
    "httpx>=0.28.1",
    "jwcrypto>=1.5.6",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.18",
    "langchain-community>=0.3.25",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.28",
    "langgraph>=0.4.8",
    "litellm>=1.73.1",
    "networkx>=3.5",
    "openstudio>=3.10.0",
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "uvicorn>=0.34.3",
]
packages = [{include = "automa_ai", from = "."}]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
